requirements.txt
requests
fpdf
python-dotenv